requirements.txt
pg8000==1.29.8
python-dotenv==1.0.0
requests==2.31.0 